Disable skipperships
