+++ .pre-commit-config.yaml
@@ -12,7 +12,7 @@ repos:
       - id: debug-statements
       - id: name-tests-test
         args: [--django]
-        exclude: .*/__mocks__.py
+        exclude: (.*/__mocks__.py|.*/__disabled__\..*.py)
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.14.6
