+++ loefsys/members/__disabled__.admin.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+
+from .models.skippership import Skippership
+from .models.user_skippership import UserSkippership
+
+
+@admin.register(Skippership, UserSkippership)
+class SkippershipAdmin(admin.ModelAdmin):
+    """Admin class for the Skippership models."""
