+++ loefsys/members/admin.py
@@ -5,8 +5,6 @@ from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
 from django.utils.translation import gettext_lazy as _
 
 from .models import User
-from .models.skippership import Skippership
-from .models.user_skippership import UserSkippership
 
 
 @admin.register(User)
@@ -51,8 +49,3 @@ class UserAdmin(BaseUserAdmin):
         if not obj:
             return self.add_fieldsets
         return super().get_fieldsets(request, obj)
-
-
-@admin.register(Skippership, UserSkippership)
-class SkippershipAdmin(admin.ModelAdmin):
-    """Admin class for the Skippership models."""
